moved dexie types to own file, added missing types and created a few live querys
